Remove custom RequestHandler implementation example.
