+++ examples/powertools-examples-kafka/template.yaml
@@ -11,30 +11,6 @@ Globals:
     Tracing: Active
 
 Resources:
-  NativeKafkaJsonConsumerDeserializationFunction:
-    Type: AWS::Serverless::Function
-    Properties:
-      CodeUri: .
-      Handler: org.demo.kafka.NativeKafkaJsonConsumerDeserializationFunction::handleRequest
-      Environment:
-        Variables:
-          JAVA_TOOL_OPTIONS: "-XX:+TieredCompilation -XX:TieredStopAtLevel=1"
-          POWERTOOLS_LOG_LEVEL: DEBUG
-          POWERTOOLS_SERVICE_NAME: NativeKafkaJsonConsumerDeserialization
-          POWERTOOLS_METRICS_NAMESPACE: NativeKafkaJsonConsumerDeserialization
-
-  NativeKafkaAvroConsumerDeserializationFunction:
-    Type: AWS::Serverless::Function
-    Properties:
-      CodeUri: .
-      Handler: org.demo.kafka.NativeKafkaAvroConsumerDeserializationFunction::handleRequest
-      Environment:
-        Variables:
-          JAVA_TOOL_OPTIONS: "-XX:+TieredCompilation -XX:TieredStopAtLevel=1"
-          POWERTOOLS_LOG_LEVEL: DEBUG
-          POWERTOOLS_SERVICE_NAME: NativeKafkaAvroConsumerDeserialization
-          POWERTOOLS_METRICS_NAMESPACE: NativeKafkaAvroConsumerDeserialization
-
   KafkaJsonConsumerDeserializationFunction:
     Type: AWS::Serverless::Function
     Properties:
