+++ powertools-kafka/src/main/java/software/amazon/lambda/powertools/kafka/serializers/KafkaJsonDeserializer.java
@@ -22,6 +22,7 @@ public class KafkaJsonDeserializer extends AbstractKafkaDeserializer {
     @Override
     protected <T> T deserializeComplex(byte[] data, Class<T> type) throws IOException {
         String decodedStr = new String(data);
+        
         return objectMapper.readValue(decodedStr, type);
     }
 }
